moved met and HadTop out from the ttHLepEventAnalyzer; added QD discrimiator tot he jets
