+++ CMGTools/TTHAnalysis/python/analyzers/ttHTopoVarAnalyzer.py
@@ -52,7 +52,38 @@ class ttHTopoVarAnalyzer( Analyzer ):
         self.counters.addCounter('pairs')
         count = self.counters.counter('pairs')
         count.register('all events')
-        
+
+    def makeHadTopDecays(self, event):
+
+        print 'Nlight',len(event.lightJets)
+
+        event.lightJets = [ j for j in event.cleanJets if not j.btagWP("CSVM") ]
+        event.minMWjj   = 999
+        event.minMWjjPt = 0
+        event.bestMWjj   = 0
+        event.bestMWjjPt = 0
+        event.bestMTopHad   = 0
+        event.bestMTopHadPt = 0
+        for i1,j1 in enumerate(event.lightJets):
+            for i2 in xrange(i1+1,len(event.lightJets)):
+                j2 = event.lightJets[i2]
+                jjp4 = j1.p4() + j2.p4()
+                mjj  = jjp4.M()
+                if mjj > 30 and mjj < event.minMWjj:
+                    event.minMWjj = mjj
+                    event.minMWjjPt = jjp4.Pt()
+                if abs(mjj-80.4) < abs(event.bestMWjj-80.4):
+                    event.bestMWjj = mjj
+                    event.bestMWjjPt = jjp4.Pt()
+                    for bj in event.bjetsLoose:
+                        if deltaR(bj.eta(),bj.phi(),j1.eta(),j1.phi()) < 0.1 or deltaR(bj.eta(),bj.phi(),j2.eta(),j2.phi()) < 0.1: continue
+                        tp4 = jjp4 + bj.p4()
+                        mtop = tp4.M()
+                        if abs(mtop-172) < abs(event.bestMTopHad - 172):
+                            event.bestMTopHad = mtop
+                            event.bestMTopHadPt = tp4.Pt()
+                            
+                                
     def printMT(self, event):
 #        print '==> INSIDE THE PRINT MT'
         event.met = self.handles['met'].product()[0]
@@ -124,6 +155,7 @@ class ttHTopoVarAnalyzer( Analyzer ):
 
         print 'variables computed: MT=',event.mtw,'MT2=',event.mt2,'MT2W=',event.mt2w
 
+        self.makeHadTopDecays(event)
 
 #        self.printMT2(event)
         return True
